refactor(analytics): extract a DripsSetEventService
