+++ src/Analytics/AccountService.ts
@@ -1,39 +1,30 @@
-/* eslint-disable no-await-in-loop */
-import type { DripsSetEvent } from 'src/DripsSubgraph/types';
 import Utils from '../utils';
 import constants from '../constants';
-import DripsSubgraphClient from '../DripsSubgraph/DripsSubgraphClient';
 import type { DripsSetEventWithFullReceivers } from './AccountEstimator/types';
 import type { Account, AssetConfig, AssetConfigHistoryItem, Stream } from './common/types';
+import DripsSetEventService from './DripsSetEventService';
 
-export interface IAccountService {
-	fetchAccount(userId: string, chainId: number): Promise<Account>;
-}
+const defaultDependencyFactory = (chainId: number): DripsSetEventService => new DripsSetEventService(chainId);
 
-const defaultDependencyFactory = (chainId: number) => DripsSubgraphClient.create(chainId);
-
-export default class AccountService implements IAccountService {
+export default class AccountService {
 	#chainId: number;
+	#dripsSetEventService: DripsSetEventService;
+
 	get chainId() {
 		return this.#chainId;
 	}
 
-	#subgraphClient: DripsSubgraphClient;
-	get subgraphClient() {
-		return this.#subgraphClient;
-	}
-
 	public constructor(
 		chainId: number,
-		dependencyFactory: (chainId: number) => DripsSubgraphClient = defaultDependencyFactory
+		dependencyFactory: (chainId: number) => DripsSetEventService = defaultDependencyFactory
 	) {
 		this.#chainId = chainId;
-		this.#subgraphClient = dependencyFactory(chainId);
+		this.#dripsSetEventService = dependencyFactory(chainId);
 	}
 
 	async fetchAccount(userId: string, chainId: number): Promise<Account> {
 		try {
-			const dripsSetEvents = await this.#getDripsSetEvents(userId);
+			const dripsSetEvents = await this.#dripsSetEventService.getAllDripsSetEvents(userId);
 			if (!userId) {
 				throw new Error(`Could fetch account: user ID is required.`);
 			}
@@ -42,8 +33,10 @@ export default class AccountService implements IAccountService {
 				throw new Error(`Could fetch account: chain ID is required.`);
 			}
 
-			const dripsSetEventsWithFullReceivers = this.#reconcileDripsSetReceivers(dripsSetEvents);
-			const dripsSetEventsByTokenAddress = this.#separateDripsSetEvents(dripsSetEventsWithFullReceivers);
+			const dripsSetEventsWithFullReceivers = this.#dripsSetEventService.reconcileDripsSetReceivers(dripsSetEvents);
+			const dripsSetEventsByTokenAddress = this.#dripsSetEventService.separateDripsSetEvents(
+				dripsSetEventsWithFullReceivers
+			);
 			const assetConfigs = this.#buildAssetConfigsForUser(userId, dripsSetEventsByTokenAddress) || [];
 
 			return {
@@ -53,90 +46,6 @@ export default class AccountService implements IAccountService {
 		} catch (error: any) {
 			throw new Error(`Could not fetch account: ${error.message}`);
 		}
-	}
-
-	async #getDripsSetEvents(userId: string) {
-		let skip = 0;
-		const first = 500;
-		const dripsSetEvents: DripsSetEvent[] = [];
-		while (true) {
-			const iterationEvents = await this.#subgraphClient.getDripsSetEventsByUserId(userId, skip, first);
-
-			dripsSetEvents.push(...iterationEvents);
-
-			if (!iterationEvents?.length || iterationEvents.length < first) {
-				break;
-			}
-
-			skip += first;
-		}
-		return dripsSetEvents;
-	}
-
-	#reconcileDripsSetReceivers(dripsSetEvents: DripsSetEvent[]): DripsSetEventWithFullReceivers[] {
-		const sortedDripsSetEvents = this.#sortDripsSetEvents(dripsSetEvents);
-
-		interface DripsReceiverSeenEvent {
-			id: string;
-			receiverUserId: string;
-			config: bigint;
-		}
-
-		const receiversHashes = sortedDripsSetEvents.reduce<string[]>((acc, dripsSetEvent) => {
-			const { receiversHash } = dripsSetEvent;
-
-			return !acc.includes(receiversHash) ? [...acc, receiversHash] : acc;
-		}, []);
-
-		const dripsReceiverSeenEventsByReceiversHash = receiversHashes.reduce<{
-			[receiversHash: string]: DripsReceiverSeenEvent[];
-		}>((acc, receiversHash) => {
-			const receivers = this.#deduplicateArray(
-				sortedDripsSetEvents
-					.filter((event) => event.receiversHash === receiversHash)
-					.reduce<DripsReceiverSeenEvent[]>((accc, event) => [...accc, ...event.dripsReceiverSeenEvents], []),
-				'config'
-			);
-
-			return {
-				...acc,
-				[receiversHash]: receivers
-			};
-		}, {});
-
-		return sortedDripsSetEvents.reduce<DripsSetEventWithFullReceivers[]>(
-			(acc, dripsSetEvent) => [
-				...acc,
-				{
-					...dripsSetEvent,
-					currentReceivers: dripsReceiverSeenEventsByReceiversHash[dripsSetEvent.receiversHash] ?? []
-				}
-			],
-			[]
-		);
-	}
-
-	#separateDripsSetEvents<T extends DripsSetEvent>(
-		dripsSetEvents: T[]
-	): {
-		[tokenAddress: string]: T[];
-	} {
-		const sorted = this.#sortDripsSetEvents(dripsSetEvents);
-
-		const result = sorted.reduce<{ [tokenAddress: string]: T[] }>((acc, dripsSetEvent) => {
-			const { assetId } = dripsSetEvent;
-			const tokenAddress = Utils.Asset.getAddressFromId(assetId);
-
-			if (acc[tokenAddress]) {
-				acc[tokenAddress].push(dripsSetEvent);
-			} else {
-				acc[tokenAddress] = [dripsSetEvent];
-			}
-
-			return acc;
-		}, {});
-
-		return result;
 	}
 
 	#buildAssetConfigsForUser(
@@ -197,13 +106,5 @@ export default class AccountService implements IAccountService {
 
 			return acc;
 		}, []);
-	}
-
-	#sortDripsSetEvents<T extends DripsSetEvent>(dripsSetEvents: T[]): T[] {
-		return dripsSetEvents.sort((a, b) => Number(a.blockTimestamp) - Number(b.blockTimestamp));
-	}
-
-	#deduplicateArray<T>(array: T[], key: keyof T): T[] {
-		return [...new Map(array.map((item) => [item[key], item])).values()];
 	}
 }
