+++ src/Analytics/ReceiverStats/types.ts
@@ -1,0 +1,1 @@
+export interface Supporter {}
